把全局的 flatf 重命名为 flat，并将参数类型从 float 改为 CGFloat，从而修复 UIImage 绘制图片时由于精度不足导致可能尺寸不正确的 bug

+++ qmuidemo/Modules/Common/Utils/UIImageEffects.m
@@ -236,7 +236,7 @@
             NSUInteger matrixSize = sizeof(floatingPointSaturationMatrix)/sizeof(floatingPointSaturationMatrix[0]);
             int16_t saturationMatrix[matrixSize];
             for (NSUInteger i = 0; i < matrixSize; ++i) {
-                saturationMatrix[i] = (int16_t)roundf(floatingPointSaturationMatrix[i] * divisor);
+                saturationMatrix[i] = (int16_t)round(floatingPointSaturationMatrix[i] * divisor);
             }
             vImageMatrixMultiply_ARGB8888(inputBuffer, outputBuffer, saturationMatrix, divisor, NULL, NULL, kvImageNoFlags);
